backend/requirements/base.txt
#Django
Django==3.2.14
django-filter==22.1
django-storages==1.12.3
djangorestframework==3.13.1

#Other
sendgrid==6.9.7
boto3==1.24.23
psycopg2==2.9.3
pycodestyle==2.8.0
Pillow==9.2.0
requests==2.28.1
django-cors-headers==3.13.0
six==1.16.0
geocoder==1.38.1
drf-nested-routers==0.93.4
geoip2==4.6.0
humanize==4.2.3
timezonefinder==6.0.1
gunicorn==20.1.0
gevent==21.12.0
pandas
django_apscheduler
numpy
sendgrid
boto3
pycodestyle
requests
geocoder
facebook-sdk
instaloader
instagrapi